cleans code
